PMM-15326: Split the OM plugin's hooks file and its callers
Three review notes on #5817 that are all about shape rather than
behaviour.

`hooks.ts` was 362 lines of which only four exports were hooks - the
rest was the fetch client, the error class, a status predicate, and the
four pure roll-ups the pages render. The inventory side already had the
split this was missing, `inventory.ts` beside `inventoryHooks.ts`, so
the topology side now matches it:

  api.ts             OmApiError, request, isRunActive - shared
  topology.ts        toServiceRows / toClusterRows / toEnvironmentSections
  topologyHooks.ts   the four topology hooks

`api.ts` is where the transport belongs because both halves use it: the
topology document and SEP's proxied estate are very different sources
that arrive over one origin with one auth story and one error envelope.
Naming the file pair `topologyHooks.ts` / `inventoryHooks.ts` also
answers the question the old name invited, which is which of the two a
given hook reads.

`useRefreshInventory` now returns `refreshAll()` and `refreshHosts(ids)`
beside the mutation state. TanStack's `mutate` is
`(variables, options?)` with `variables` positional, so the old
`string[] | undefined` variable forced every unscoped caller to write
`mutate(undefined)` - which reads as an oversight and had to be
explained twice in review. Three call sites now say which sweep they
are asking for.

Components are arrow functions, matching `apps/pmm`, which is 53 arrow
declarations to none. 30 components across 13 files; hooks and plain
helpers keep their function declarations.

The roll-up's `!= null` guards gained a comment on the way across,
because they are the ones that must not become `!== null`: those fields
are proto3 `optional` scalars, which protojson omits entirely rather
than nulling, so the absent case is undefined and a strict guard would
put it into Math.max and report NaN for the cluster.
